# docs/swagger.yaml
basePath: /
definitions:
...
  tax-priority-api_src_presentation_models.BatchCommandResult:
    properties:
      errors:
        example:
        - '["Validation failed for item 1"]'
        items:
          type: string
        type: array
      failureCount:
        example: 2
        type: integer
      results:
        items:
          $ref: '#/definitions/tax-priority-api_src_presentation_models.CommandResult'
        type: array
      successCount:
        example: 8
        type: integer
    type: object
  tax-priority-api_src_presentation_models.BulkDeleteFAQRequest:
    properties:
      ids:
        example:
        - '["uuid1"'
        - ' "uuid2"]'
        items:
          type: string
        minItems: 1
        type: array
    required:
    - ids
    type: object
  tax-priority-api_src_presentation_models.CommandResult:
    properties:
      createdAt:
        example: "2023-12-01T10:00:00Z"
        type: string
      error:
        example: Validation failed
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      message:
        example: FAQ created successfully
        type: string
      success:
        example: true
        type: boolean
      updatedAt:
        example: "2023-12-01T10:00:00Z"
        type: string
    type: object
  tax-priority-api_src_presentation_models.CountResponse:
    properties:
      count:
        example: 42
        type: integer
    type: object
  tax-priority-api_src_presentation_models.ErrorResponse:
    properties:
      error:
        example: Validation failed
        type: string
    type: object
  tax-priority-api_src_presentation_models.FAQResponse:
    properties:
      answer:
        example: Для подачи налоговой декларации необходимо...
        type: string
      category:
        example: налоги
        type: string
      createdAt:
        example: "2023-12-01T10:00:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      isActive:
        example: true
        type: boolean
      priority:
        example: 50
        type: integer
      question:
        example: Как подать налоговую декларацию?
        type: string
      updatedAt:
        example: "2023-12-01T10:00:00Z"
        type: string
    type: object
  tax-priority-api_src_presentation_models.GetFAQsByIDsRequest:
    properties:
      ids:
        example:
        - '["uuid1"'
        - ' "uuid2"]'
        items:
          type: string
        minItems: 1
        type: array
    required:
    - ids
    type: object
  tax-priority-api_src_presentation_models.PaginatedFAQResponse:
    properties:
      hasNext:
        example: true
        type: boolean
      hasPrev:
        example: false
        type: boolean
      items:
        items:
          $ref: '#/definitions/tax-priority-api_src_presentation_models.FAQResponse'
        type: array
      limit:
        example: 10
        type: integer
      offset:
        example: 0
        type: integer
      total:
        example: 100
        type: integer
      totalPages:
        example: 10
        type: integer
    type: object
  tax-priority-api_src_presentation_models.UpdateFAQPriorityRequest:
    properties:
      priority:
        example: 75
        maximum: 100
        minimum: 0
        type: integer
    type: object
  tax-priority-api_src_presentation_models.UpdateFAQRequest:
    properties:
      answer:
        example: Для подачи налоговой декларации необходимо...
        maxLength: 2000
        minLength: 10
        type: string
      category:
        example: налоги
        maxLength: 100
        type: string
      priority:
        example: 50
        maximum: 100
        minimum: 0
        type: integer
      question:
        example: Как подать налоговую декларацию?
        maxLength: 500
        minLength: 10
        type: string
    required:
    - answer
    - category
    - question
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API для управления FAQ в системе Tax Priority
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Tax Priority API
  version: "1.0"
paths:
  /api/faq:
    get:
      description: Возвращает список FAQ с пагинацией и фильтрацией
      parameters:
      - default: 10
        description: Лимит записей
        in: query
        name: _limit
        type: integer
      - default: 0
        description: Смещение
        in: query
        name: _offset
        type: integer
      - default: createdAt
        description: Поле сортировки
        in: query
        name: _sort
        type: string
      - default: desc
        description: Порядок сортировки
        enum:
        - asc
        - desc
        in: query
        name: _order
        type: string
      - description: Фильтр по категории
        in: query
        name: category
        type: string
      - description: Фильтр по активности
        in: query
        name: isActive
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.PaginatedFAQResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
      summary: Получить список FAQ
      tags:
      - FAQ
  /api/faq/{id}:
    delete:
      description: Удаляет FAQ по ID
      parameters:
      - description: ID FAQ
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.CommandResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
      summary: Удалить FAQ
      tags:
      - FAQ
    get:
      description: Возвращает FAQ по указанному ID
      parameters:
      - description: ID FAQ
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.FAQResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
      summary: Получить FAQ по ID
      tags:
      - FAQ
    put:
      consumes:
      - application/json
      description: Обновляет существующую FAQ
      parameters:
      - description: ID FAQ
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: faq
        required: true
        schema:
          $ref: '#/definitions/tax-priority-api_src_presentation_models.UpdateFAQRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.CommandResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
      summary: Обновить FAQ
      tags:
      - FAQ
  /api/faq/{id}/activate:
    patch:
      description: Активирует FAQ по ID
      parameters:
      - description: ID FAQ
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.CommandResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
      summary: Активировать FAQ
      tags:
      - FAQ
  /api/faq/{id}/deactivate:
    patch:
      description: Деактивирует FAQ по ID
      parameters:
      - description: ID FAQ
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.CommandResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
      summary: Деактивировать FAQ
      tags:
      - FAQ
  /api/faq/{id}/priority:
    patch:
      consumes:
      - application/json
      description: Обновляет приоритет FAQ по ID
      parameters:
      - description: ID FAQ
        in: path
        name: id
        required: true
        type: string
      - description: Новый приоритет
        in: body
        name: priority
        required: true
        schema:
          $ref: '#/definitions/tax-priority-api_src_presentation_models.UpdateFAQPriorityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.CommandResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
      summary: Обновить приоритет FAQ
      tags:
      - FAQ
  /api/faq/batch:
    post:
      consumes:
      - application/json
      description: Возвращает FAQ по списку ID (batch запрос)
      parameters:
      - description: Список ID
        in: body
        name: ids
        required: true
        schema:
          $ref: '#/definitions/tax-priority-api_src_presentation_models.GetFAQsByIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tax-priority-api_src_presentation_models.FAQResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
      summary: Получить FAQ по списку ID
      tags:
      - FAQ
  /api/faq/bulk-delete:
    delete:
      consumes:
      - application/json
      description: Удаляет несколько FAQ по списку ID
      parameters:
      - description: Список ID для удаления
        in: body
        name: ids
        required: true
        schema:
          $ref: '#/definitions/tax-priority-api_src_presentation_models.BulkDeleteFAQRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.BatchCommandResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
      summary: Массовое удаление FAQ
      tags:
      - FAQ
  /api/faq/count:
    get:
      description: Возвращает общее количество FAQ
      parameters:
      - description: Фильтр по категории
        in: query
        name: category
        type: string
      - description: Фильтр по активности
        in: query
        name: isActive
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.CountResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/tax-priority-api_src_presentation_models.ErrorResponse'
      summary: Получить количество FAQ
      tags:
      - FAQ
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"